==> 00_scripts/17_glm_with_iterations/28_find_significant_SNP.R <==
#!/usr/bin/env Rscript

setwd("/home1/datawork/plstenge/Pearl_Oyster_Colour_Population_Genomics/08_02_input_pvalued_without_iteration")

dat <- read.table("all_glm_tuckey_without_iteration_results.txt", header=F, fill=TRUE)
colnames(dat) <- c("Ten", "SNP", "RedVsGreen", "YellowVsGreen", "YellowVsRed", "HatcheryVsGambier", "KatiuVsGambier", "TakapotoVsGambier", "KatiuVsHatchery","TakapotoVsHatchery", "TakapotoVsKatiu")
dat$RedVsGreen <- as.numeric(as.character(dat$RedVsGreen ))
dat$RedVsGreen <- as.numeric(as.character(dat$RedVsGreen ))
dat$YellowVsGreen <- as.numeric(as.character(dat$YellowVsGreen ))
dat$YellowVsGreen <- as.numeric(as.character(dat$YellowVsGreen ))
dat$YellowVsRed <- as.numeric(as.character(dat$YellowVsRed ))
dat$YellowVsRed <- as.numeric(as.character(dat$YellowVsRed ))
dat$HatcheryVsGambier <- as.numeric(as.character(dat$HatcheryVsGambier ))
dat$HatcheryVsGambier <- as.numeric(as.character(dat$HatcheryVsGambier ))
dat$KatiuVsGambier <- as.numeric(as.character(dat$KatiuVsGambier ))
dat$KatiuVsGambier <- as.numeric(as.character(dat$KatiuVsGambier ))
dat$TakapotoVsGambier <- as.numeric(as.character(dat$TakapotoVsGambier ))
dat$TakapotoVsGambier <- as.numeric(as.character(dat$TakapotoVsGambier ))
dat$KatiuVsHatchery <- as.numeric(as.character(dat$KatiuVsHatchery ))
dat$KatiuVsHatchery <- as.numeric(as.character(dat$KatiuVsHatchery ))
dat$TakapotoVsHatchery <- as.numeric(as.character(dat$TakapotoVsHatchery ))
dat$TakapotoVsHatchery <- as.numeric(as.character(dat$TakapotoVsHatchery ))
dat$TakapotoVsKatiu <- as.numeric(as.character(dat$TakapotoVsKatiu ))
dat$TakapotoVsKatiu <- as.numeric(as.character(dat$TakapotoVsKatiu ))
dat$RedVsGreen <- as.numeric(dat$RedVsGreen)
print("str((dat$RedVsGreen))")
str(dat$RedVsGreen)
max(dat$RedVsGreen)
min(dat$RedVsGreen)
datRedVsGreen <- subset(dat, RedVsGreen<0.000000001)
write.table(datRedVsGreen, "RedVsGreen_dat_P_0_000000001.txt")
datRedVsGreen <- subset(dat, RedVsGreen<0.000001)
write.table(datRedVsGreen, "RedVsGreen_dat_P_0_000001.txt")
datRedVsGreen <- subset(dat, RedVsGreen<0.001)
write.table(datRedVsGreen, "RedVsGreen_dat_P_0_001.txt")
datRedVsGreen <- subset(dat, RedVsGreen<0.01)
write.table(datRedVsGreen, "RedVsGreen_dat_P_0_01.txt")
dat$YellowVsGreen <- as.numeric(dat$YellowVsGreen)
print("str((dat$YellowVsGreen))")
str(dat$YellowVsGreen)
max(dat$YellowVsGreen)
min(dat$YellowVsGreen)
datYellowVsGreen <- subset(dat, YellowVsGreen<0.000000001)
write.table(datYellowVsGreen, "YellowVsGreen_dat_P_0_000000001.txt")
datYellowVsGreen <- subset(dat, YellowVsGreen<0.000001)
write.table(datYellowVsGreen, "YellowVsGreen_dat_P_0_000001.txt")
datYellowVsGreen <- subset(dat, YellowVsGreen<0.001)
write.table(datYellowVsGreen, "YellowVsGreen_dat_P_0_001.txt")
datYellowVsGreen <- subset(dat, YellowVsGreen<0.01)
write.table(datYellowVsGreen, "YellowVsGreen_dat_P_0_01.txt")
dat$YellowVsRed <- as.numeric(dat$YellowVsRed)
print("str((dat$YellowVsRed))")
str(dat$YellowVsRed)
max(dat$YellowVsRed)
min(dat$YellowVsRed)
datYellowVsRed <- subset(dat, YellowVsRed<0.000000001)
write.table(datYellowVsRed, "YellowVsRed_dat_P_0_000000001.txt")
datYellowVsRed <- subset(dat, YellowVsRed<0.000001)
write.table(datYellowVsRed, "YellowVsRed_dat_P_0_000001.txt")
datYellowVsRed <- subset(dat, YellowVsRed<0.001)
write.table(datYellowVsRed, "YellowVsRed_dat_P_0_001.txt")
datYellowVsRed <- subset(dat, YellowVsRed<0.01)
write.table(datYellowVsRed, "YellowVsRed_dat_P_0_01.txt")
dat$HatcheryVsGambier <- as.numeric(dat$HatcheryVsGambier)
print("str((dat$HatcheryVsGambier))")
str(dat$HatcheryVsGambier)
max(dat$HatcheryVsGambier)
min(dat$HatcheryVsGambier)
datHatcheryVsGambier <- subset(dat, HatcheryVsGambier<0.000000001)
write.table(datHatcheryVsGambier, "HatcheryVsGambier_dat_P_0_000000001.txt")
datHatcheryVsGambier <- subset(dat, HatcheryVsGambier<0.000001)
write.table(datHatcheryVsGambier, "HatcheryVsGambier_dat_P_0_000001.txt")
datHatcheryVsGambier <- subset(dat, HatcheryVsGambier<0.001)
write.table(datHatcheryVsGambier, "HatcheryVsGambier_dat_P_0_001.txt")
datHatcheryVsGambier <- subset(dat, HatcheryVsGambier<0.01)
write.table(datHatcheryVsGambier, "HatcheryVsGambier_dat_P_0_01.txt")
dat$KatiuVsGambier <- as.numeric(dat$KatiuVsGambier)
print("str((dat$KatiuVsGambier))")
str(dat$KatiuVsGambier)
max(dat$KatiuVsGambier)
min(dat$KatiuVsGambier)
datKatiuVsGambier <- subset(dat, KatiuVsGambier<0.000000001)
write.table(datKatiuVsGambier, "KatiuVsGambier_dat_P_0_000000001.txt")
datKatiuVsGambier <- subset(dat, KatiuVsGambier<0.000001)
write.table(datKatiuVsGambier, "KatiuVsGambier_dat_P_0_000001.txt")
datKatiuVsGambier <- subset(dat, KatiuVsGambier<0.001)
write.table(datKatiuVsGambier, "KatiuVsGambier_dat_P_0_001.txt")
datKatiuVsGambier <- subset(dat, KatiuVsGambier<0.01)
write.table(datKatiuVsGambier, "KatiuVsGambier_dat_P_0_01.txt")
dat$TakapotoVsGambier <- as.numeric(dat$TakapotoVsGambier)
print("str((dat$TakapotoVsGambier))")
str(dat$TakapotoVsGambier)
max(dat$TakapotoVsGambier)
min(dat$TakapotoVsGambier)
datTakapotoVsGambier <- subset(dat, TakapotoVsGambier<0.000000001)
write.table(datTakapotoVsGambier, "TakapotoVsGambier_dat_P_0_000000001.txt")
datTakapotoVsGambier <- subset(dat, TakapotoVsGambier<0.000001)
write.table(datTakapotoVsGambier, "TakapotoVsGambier_dat_P_0_000001.txt")
datTakapotoVsGambier <- subset(dat, TakapotoVsGambier<0.001)
write.table(datTakapotoVsGambier, "TakapotoVsGambier_dat_P_0_001.txt")
datTakapotoVsGambier <- subset(dat, TakapotoVsGambier<0.01)
write.table(datTakapotoVsGambier, "TakapotoVsGambier_dat_P_0_01.txt")
dat$KatiuVsHatchery <- as.numeric(dat$KatiuVsHatchery)
print("str((dat$KatiuVsHatchery))")
str(dat$KatiuVsHatchery)
max(dat$KatiuVsHatchery)
min(dat$KatiuVsHatchery)
datKatiuVsHatchery <- subset(dat, KatiuVsHatchery<0.000000001)
write.table(datKatiuVsHatchery, "KatiuVsHatchery_dat_P_0_000000001.txt")
datKatiuVsHatchery <- subset(dat, KatiuVsHatchery<0.000001)
write.table(datKatiuVsHatchery, "KatiuVsHatchery_dat_P_0_000001.txt")
datKatiuVsHatchery <- subset(dat, KatiuVsHatchery<0.001)
write.table(datKatiuVsHatchery, "KatiuVsHatchery_dat_P_0_001.txt")
datKatiuVsHatchery <- subset(dat, KatiuVsHatchery<0.01)
write.table(datKatiuVsHatchery, "KatiuVsHatchery_dat_P_0_01.txt")
dat$TakapotoVsHatchery <- as.numeric(dat$TakapotoVsHatchery)
print("str((dat$TakapotoVsHatchery))")
str(dat$TakapotoVsHatchery)
max(dat$TakapotoVsHatchery)
min(dat$TakapotoVsHatchery)
datTakapotoVsHatchery <- subset(dat, TakapotoVsHatchery<0.000000001)
write.table(datTakapotoVsHatchery, "TakapotoVsHatchery_dat_P_0_000000001.txt")
datTakapotoVsHatchery <- subset(dat, TakapotoVsHatchery<0.000001)
write.table(datTakapotoVsHatchery, "TakapotoVsHatchery_dat_P_0_000001.txt")
datTakapotoVsHatchery <- subset(dat, TakapotoVsHatchery<0.001)
write.table(datTakapotoVsHatchery, "TakapotoVsHatchery_dat_P_0_001.txt")
datTakapotoVsHatchery <- subset(dat, TakapotoVsHatchery<0.01)
write.table(datTakapotoVsHatchery, "TakapotoVsHatchery_dat_P_0_01.txt")
dat$TakapotoVsKatiu <- as.numeric(dat$TakapotoVsKatiu)
print("str((dat$TakapotoVsKatiu))")
str(dat$TakapotoVsKatiu)
max(dat$TakapotoVsKatiu)
min(dat$TakapotoVsKatiu)
datTakapotoVsKatiu <- subset(dat, TakapotoVsKatiu<0.000000001)
write.table(datTakapotoVsKatiu, "TakapotoVsKatiu_dat_P_0_000000001.txt")
datTakapotoVsKatiu <- subset(dat, TakapotoVsKatiu<0.000001)
write.table(datTakapotoVsKatiu, "TakapotoVsKatiu_dat_P_0_000001.txt")
datTakapotoVsKatiu <- subset(dat, TakapotoVsKatiu<0.001)
write.table(datTakapotoVsKatiu, "TakapotoVsKatiu_dat_P_0_001.txt")
datTakapotoVsKatiu <- subset(dat, TakapotoVsKatiu<0.01)
write.table(datTakapotoVsKatiu, "TakapotoVsKatiu_dat_P_0_01.txt")
