==> 00_scripts/17_glm_with_iterations/28_02_chord_diagramm.R <==
#!/usr/bin/env Rscript


#######################################################################################################################
# Set Working Directory
#######################################################################################################################

setwd("/home1/datawork/plstenge/Pearl_Oyster_Colour_Population_Genomics/08_02_input_pvalued_without_iteration")

#######################################################################################################################
# Import data
#######################################################################################################################

HatcheryVsGambier<-read.table("HatcheryVsGambier_dat_P_0_000000001.txt_SNP3.txt", header=T)
KatiuVsGambier<-read.table("KatiuVsGambier_dat_P_0_000000001.txt_SNP3.txt", header=T)
KatiuVsHatchery<-read.table("KatiuVsHatchery_dat_P_0_000000001.txt_SNP3.txt", header=T)
RedVsGreen<-read.table("RedVsGreen_dat_P_0_000000001.txt_SNP3.txt", header=T)
TakapotoVsGambier<-read.table("TakapotoVsGambier_dat_P_0_000000001.txt_SNP3.txt", header=T)
TakapotoVsHatchery<-read.table("TakapotoVsHatchery_dat_P_0_000000001.txt_SNP3.txt", header=T)
TakapotoVsKatiu<-read.table("TakapotoVsKatiu_dat_P_0_000000001.txt_SNP3.txt", header=T)
YellowVsGreen<-read.table("YellowVsGreen_dat_P_0_000000001.txt_SNP3.txt", header=T)
YellowVsRed<-read.table("YellowVsRed_dat_P_0_000000001.txt_SNP3.txt", header=T)

print("#######################################################")
print("P_0_000000001")
print("#######################################################")

commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsGambier$RedVsGreen)
head(commun)
str(commun)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSKatiuVsGambier.txt")
print(paste0("HatcheryVsGambier, KatiuVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSKatiuVsHatchery.txt")
print(paste0("HatcheryVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSRedVsGreen.txt")
print(paste0("HatcheryVsGambier, RedVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSTakapotoVsGambier.txt")
print(paste0("HatcheryVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSTakapotoVsHatchery.txt")
print(paste0("HatcheryVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSTakapotoVsKatiu.txt")
print(paste0("HatcheryVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSellowVsGreen.txt")
print(paste0("HatcheryVsGambier, YellowVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_HatcheryVsGambierVSYellowVsRed.txt")
print(paste0("HatcheryVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSKatiuVsHatchery.txt")
print(paste0("KatiuVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSRedVsGreen.txt")
print(paste0("KatiuVsGambier, RedVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSTakapotoVsGambier.txt")
print(paste0("KatiuVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSYellowVsGreen.txt")
print(paste0("KatiuVsGambier, YellowVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsGambierVSYellowVsRed.txt")
print(paste0("KatiuVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSedVsGreen.txt")
print(paste0("KatiuVsHatchery, RedVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSTakapotoVsGambier.txt")
print(paste0("KatiuVsHatchery, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsHatchery, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSellowVsGreen.txt")
print(paste0("KatiuVsHatchery, YellowVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_KatiuVsHatcheryVSYellowVsRed.txt")
print(paste0("KatiuVsHatchery, YellowVsRed	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_RedVsGreenVSakapotoVsGambier.txt")
print(paste0("RedVsGreen, TakapotoVsGambier	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_RedVsGreenVSTakapotoVsHatchery.txt")
print(paste0("RedVsGreen, TakapotoVsHatchery	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_RedVsGreenVSTakapotoVsKatiu.txt")
print(paste0("RedVsGreen, TakapotoVsKatiu	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_RedVsGreenVSYellowVsGreen.txt")
print(paste0("RedVsGreen, YellowVsGreen	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_RedVsGreenVSYellowVsRed.txt")
print(paste0("RedVsGreen, YellowVsRed	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsGambierVSakapotoVsHatchery.txt")
print(paste0("TakapotoVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsGambierVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsGambierVSYellowVsGreen.txt")
print(paste0("TakapotoVsGambier, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsGambierVSYellowVsRed.txt")
print(paste0("TakapotoVsGambier, YellowVsRed	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsHatcheryVSYellowVsGreen.txt")
print(paste0("TakapotoVsHatchery, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsHatcheryVSYellowVsRed.txt")
print(paste0("TakapotoVsHatchery, YellowVsRed	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsKatiuVSellowVsGreen.txt")
print(paste0("TakapotoVsKatiu, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_TakapotoVsKatiuVSellowVsRed.txt")
print(paste0("TakapotoVsKatiu, YellowVsRed	:", res))
commun<-intersect(YellowVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000000001_YellowVsGreenVSellowVsRed.txt")
print(paste0("YellowVsGreen, YellowVsRed	:", res))

HatcheryVsGambier<-read.table("HatcheryVsGambier_dat_P_0_000001.txt_SNP3.txt", header=T)
KatiuVsGambier<-read.table("KatiuVsGambier_dat_P_0_000001.txt_SNP3.txt", header=T)
KatiuVsHatchery<-read.table("KatiuVsHatchery_dat_P_0_000001.txt_SNP3.txt", header=T)
RedVsGreen<-read.table("RedVsGreen_dat_P_0_000001.txt_SNP3.txt", header=T)
TakapotoVsGambier<-read.table("TakapotoVsGambier_dat_P_0_000001.txt_SNP3.txt", header=T)
TakapotoVsHatchery<-read.table("TakapotoVsHatchery_dat_P_0_000001.txt_SNP3.txt", header=T)
TakapotoVsKatiu<-read.table("TakapotoVsKatiu_dat_P_0_000001.txt_SNP3.txt", header=T)
YellowVsGreen<-read.table("YellowVsGreen_dat_P_0_000001.txt_SNP3.txt", header=T)
YellowVsRed<-read.table("YellowVsRed_dat_P_0_000001.txt_SNP3.txt", header=T)

print("#######################################################")
print("P_0_000001")
print("#######################################################")

commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSKatiuVsGambier.txt")
print(paste0("HatcheryVsGambier, KatiuVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSKatiuVsHatchery.txt")
print(paste0("HatcheryVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSRedVsGreen.txt")
print(paste0("HatcheryVsGambier, RedVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSTakapotoVsGambier.txt")
print(paste0("HatcheryVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSTakapotoVsHatchery.txt")
print(paste0("HatcheryVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSTakapotoVsKatiu.txt")
print(paste0("HatcheryVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSellowVsGreen.txt")
print(paste0("HatcheryVsGambier, YellowVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_HatcheryVsGambierVSYellowVsRed.txt")
print(paste0("HatcheryVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSKatiuVsHatchery.txt")
print(paste0("KatiuVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSRedVsGreen.txt")
print(paste0("KatiuVsGambier, RedVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSTakapotoVsGambier.txt")
print(paste0("KatiuVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSYellowVsGreen.txt")
print(paste0("KatiuVsGambier, YellowVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsGambierVSYellowVsRed.txt")
print(paste0("KatiuVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSedVsGreen.txt")
print(paste0("KatiuVsHatchery, RedVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSTakapotoVsGambier.txt")
print(paste0("KatiuVsHatchery, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsHatchery, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSellowVsGreen.txt")
print(paste0("KatiuVsHatchery, YellowVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_KatiuVsHatcheryVSYellowVsRed.txt")
print(paste0("KatiuVsHatchery, YellowVsRed	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_RedVsGreenVSakapotoVsGambier.txt")
print(paste0("RedVsGreen, TakapotoVsGambier	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_RedVsGreenVSTakapotoVsHatchery.txt")
print(paste0("RedVsGreen, TakapotoVsHatchery	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_RedVsGreenVSTakapotoVsKatiu.txt")
print(paste0("RedVsGreen, TakapotoVsKatiu	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_RedVsGreenVSYellowVsGreen.txt")
print(paste0("RedVsGreen, YellowVsGreen	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_RedVsGreenVSYellowVsRed.txt")
print(paste0("RedVsGreen, YellowVsRed	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsGambierVSakapotoVsHatchery.txt")
print(paste0("TakapotoVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsGambierVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsGambierVSYellowVsGreen.txt")
print(paste0("TakapotoVsGambier, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsGambierVSYellowVsRed.txt")
print(paste0("TakapotoVsGambier, YellowVsRed	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsHatcheryVSYellowVsGreen.txt")
print(paste0("TakapotoVsHatchery, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsHatcheryVSYellowVsRed.txt")
print(paste0("TakapotoVsHatchery, YellowVsRed	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsKatiuVSellowVsGreen.txt")
print(paste0("TakapotoVsKatiu, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_TakapotoVsKatiuVSellowVsRed.txt")
print(paste0("TakapotoVsKatiu, YellowVsRed	:", res))
commun<-intersect(YellowVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_000001_YellowVsGreenVSellowVsRed.txt")
print(paste0("YellowVsGreen, YellowVsRed	:", res))


HatcheryVsGambier<-read.table("HatcheryVsGambier_dat_P_0_001.txt_SNP3.txt", header=T)
KatiuVsGambier<-read.table("KatiuVsGambier_dat_P_0_001.txt_SNP3.txt", header=T)
KatiuVsHatchery<-read.table("KatiuVsHatchery_dat_P_0_001.txt_SNP3.txt", header=T)
RedVsGreen<-read.table("RedVsGreen_dat_P_0_001.txt_SNP3.txt", header=T)
TakapotoVsGambier<-read.table("TakapotoVsGambier_dat_P_0_001.txt_SNP3.txt", header=T)
TakapotoVsHatchery<-read.table("TakapotoVsHatchery_dat_P_0_001.txt_SNP3.txt", header=T)
TakapotoVsKatiu<-read.table("TakapotoVsKatiu_dat_P_0_001.txt_SNP3.txt", header=T)
YellowVsGreen<-read.table("YellowVsGreen_dat_P_0_001.txt_SNP3.txt", header=T)
YellowVsRed<-read.table("YellowVsRed_dat_P_0_001.txt_SNP3.txt", header=T)

print("#######################################################")
print("P_0_001")
print("#######################################################")

commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSKatiuVsGambier.txt")
print(paste0("HatcheryVsGambier, KatiuVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSKatiuVsHatchery.txt")
print(paste0("HatcheryVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSRedVsGreen.txt")
print(paste0("HatcheryVsGambier, RedVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSTakapotoVsGambier.txt")
print(paste0("HatcheryVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSTakapotoVsHatchery.txt")
print(paste0("HatcheryVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSTakapotoVsKatiu.txt")
print(paste0("HatcheryVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSellowVsGreen.txt")
print(paste0("HatcheryVsGambier, YellowVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_HatcheryVsGambierVSYellowVsRed.txt")
print(paste0("HatcheryVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSKatiuVsHatchery.txt")
print(paste0("KatiuVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSRedVsGreen.txt")
print(paste0("KatiuVsGambier, RedVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSTakapotoVsGambier.txt")
print(paste0("KatiuVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSYellowVsGreen.txt")
print(paste0("KatiuVsGambier, YellowVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsGambierVSYellowVsRed.txt")
print(paste0("KatiuVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSedVsGreen.txt")
print(paste0("KatiuVsHatchery, RedVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSTakapotoVsGambier.txt")
print(paste0("KatiuVsHatchery, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsHatchery, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSellowVsGreen.txt")
print(paste0("KatiuVsHatchery, YellowVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_KatiuVsHatcheryVSYellowVsRed.txt")
print(paste0("KatiuVsHatchery, YellowVsRed	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_RedVsGreenVSakapotoVsGambier.txt")
print(paste0("RedVsGreen, TakapotoVsGambier	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_RedVsGreenVSTakapotoVsHatchery.txt")
print(paste0("RedVsGreen, TakapotoVsHatchery	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_RedVsGreenVSTakapotoVsKatiu.txt")
print(paste0("RedVsGreen, TakapotoVsKatiu	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_RedVsGreenVSYellowVsGreen.txt")
print(paste0("RedVsGreen, YellowVsGreen	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_RedVsGreenVSYellowVsRed.txt")
print(paste0("RedVsGreen, YellowVsRed	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsGambierVSakapotoVsHatchery.txt")
print(paste0("TakapotoVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsGambierVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsGambierVSYellowVsGreen.txt")
print(paste0("TakapotoVsGambier, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsGambierVSYellowVsRed.txt")
print(paste0("TakapotoVsGambier, YellowVsRed	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsHatcheryVSYellowVsGreen.txt")
print(paste0("TakapotoVsHatchery, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsHatcheryVSYellowVsRed.txt")
print(paste0("TakapotoVsHatchery, YellowVsRed	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsKatiuVSellowVsGreen.txt")
print(paste0("TakapotoVsKatiu, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_TakapotoVsKatiuVSellowVsRed.txt")
print(paste0("TakapotoVsKatiu, YellowVsRed	:", res))
commun<-intersect(YellowVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_001_YellowVsGreenVSellowVsRed.txt")
print(paste0("YellowVsGreen, YellowVsRed	:", res))


HatcheryVsGambier<-read.table("HatcheryVsGambier_dat_P_0_01.txt_SNP3.txt", header=T)
KatiuVsGambier<-read.table("KatiuVsGambier_dat_P_0_01.txt_SNP3.txt", header=T)
KatiuVsHatchery<-read.table("KatiuVsHatchery_dat_P_0_01.txt_SNP3.txt", header=T)
RedVsGreen<-read.table("RedVsGreen_dat_P_0_01.txt_SNP3.txt", header=T)
TakapotoVsGambier<-read.table("TakapotoVsGambier_dat_P_0_01.txt_SNP3.txt", header=T)
TakapotoVsHatchery<-read.table("TakapotoVsHatchery_dat_P_0_01.txt_SNP3.txt", header=T)
TakapotoVsKatiu<-read.table("TakapotoVsKatiu_dat_P_0_01.txt_SNP3.txt", header=T)
YellowVsGreen<-read.table("YellowVsGreen_dat_P_0_01.txt_SNP3.txt", header=T)
YellowVsRed<-read.table("YellowVsRed_dat_P_0_01.txt_SNP3.txt", header=T)

print("#######################################################")
print("P_0_01")
print("#######################################################")

commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSKatiuVsGambier.txt")
print(paste0("HatcheryVsGambier, KatiuVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSKatiuVsHatchery.txt")
print(paste0("HatcheryVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSRedVsGreen.txt")
print(paste0("HatcheryVsGambier, RedVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSTakapotoVsGambier.txt")
print(paste0("HatcheryVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSTakapotoVsHatchery.txt")
print(paste0("HatcheryVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSTakapotoVsKatiu.txt")
print(paste0("HatcheryVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSellowVsGreen.txt")
print(paste0("HatcheryVsGambier, YellowVsGreen	:", res))
commun<-intersect(HatcheryVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_HatcheryVsGambierVSYellowVsRed.txt")
print(paste0("HatcheryVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, KatiuVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSKatiuVsHatchery.txt")
print(paste0("KatiuVsGambier, KatiuVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSRedVsGreen.txt")
print(paste0("KatiuVsGambier, RedVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSTakapotoVsGambier.txt")
print(paste0("KatiuVsGambier, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSYellowVsGreen.txt")
print(paste0("KatiuVsGambier, YellowVsGreen	:", res))
commun<-intersect(KatiuVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsGambierVSYellowVsRed.txt")
print(paste0("KatiuVsGambier, YellowVsRed	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, RedVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSedVsGreen.txt")
print(paste0("KatiuVsHatchery, RedVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSTakapotoVsGambier.txt")
print(paste0("KatiuVsHatchery, TakapotoVsGambier	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSTakapotoVsHatchery.txt")
print(paste0("KatiuVsHatchery, TakapotoVsHatchery	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("KatiuVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSellowVsGreen.txt")
print(paste0("KatiuVsHatchery, YellowVsGreen	:", res))
commun<-intersect(KatiuVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_KatiuVsHatcheryVSYellowVsRed.txt")
print(paste0("KatiuVsHatchery, YellowVsRed	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsGambier$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_RedVsGreenVSakapotoVsGambier.txt")
print(paste0("RedVsGreen, TakapotoVsGambier	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_RedVsGreenVSTakapotoVsHatchery.txt")
print(paste0("RedVsGreen, TakapotoVsHatchery	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_RedVsGreenVSTakapotoVsKatiu.txt")
print(paste0("RedVsGreen, TakapotoVsKatiu	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_RedVsGreenVSYellowVsGreen.txt")
print(paste0("RedVsGreen, YellowVsGreen	:", res))
commun<-intersect(RedVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_RedVsGreenVSYellowVsRed.txt")
print(paste0("RedVsGreen, YellowVsRed	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsHatchery$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsGambierVSakapotoVsHatchery.txt")
print(paste0("TakapotoVsGambier, TakapotoVsHatchery	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsGambierVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsGambier, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsGambierVSYellowVsGreen.txt")
print(paste0("TakapotoVsGambier, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsGambier$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsGambierVSYellowVsRed.txt")
print(paste0("TakapotoVsGambier, YellowVsRed	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, TakapotoVsKatiu$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsHatcheryVSTakapotoVsKatiu.txt")
print(paste0("TakapotoVsHatchery, TakapotoVsKatiu	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsHatcheryVSYellowVsGreen.txt")
print(paste0("TakapotoVsHatchery, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsHatchery$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsHatcheryVSYellowVsRed.txt")
print(paste0("TakapotoVsHatchery, YellowVsRed	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsGreen$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsKatiuVSellowVsGreen.txt")
print(paste0("TakapotoVsKatiu, YellowVsGreen	:", res))
commun<-intersect(TakapotoVsKatiu$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_TakapotoVsKatiuVSellowVsRed.txt")
print(paste0("TakapotoVsKatiu, YellowVsRed	:", res))
commun<-intersect(YellowVsGreen$RedVsGreen, YellowVsRed$RedVsGreen)
res <- length(commun)
write.table(commun, append = TRUE,  file = "P_0_01_YellowVsGreenVSellowVsRed.txt")
print(paste0("YellowVsGreen, YellowVsRed	:", res))
